refactoring helpers into concern, cleaning up re: PR comments
