+++ app/controllers/api/v1/metadata_schemas_controller.rb
@@ -2,12 +2,7 @@ module Api
   module V1
     class MetadataSchemasController < ApiBaseController
       include Swagger::Blocks
-      before_action :set_available_schemas, only: :index
-      before_action :set_schema_filename, only: :load_schema
-      before_action :set_schema_filepath, only: :load_schema
-      before_action :set_response_type, only: :load_schema
-
-      SCHEMAS_BASE_DIR = Rails.root.join('lib', 'assets', 'metadata_schemas')
+      include Concerns::ConventionSchemas
 
       swagger_path '/metadata_schemas' do
         operation :get do
@@ -24,7 +19,8 @@ module Api
       end
 
       def index
-        render json: @schemas
+        schemas = set_available_schemas
+        render json: schemas
       end
 
       swagger_path '/metadata_schemas/{project_name}/{version}/{schema_format}' do
@@ -33,14 +29,13 @@ module Api
               'MetadataSchemas'
           ]
           key :summary, 'Load a metadata convention schema file'
-          key :description, 'Returns a schema definition file for the requested metadata convention, in JSON or TSV format'
+          key :description, "Returns a schema definition file for the requested metadata convention, in JSON or TSV format. Refer above for available schemas."
           key :operationId, 'metadata_schemas_load_schema_path'
           parameter do
             key :name, :project_name
             key :in, :path
             key :description, 'Project name of metadata convention'
             key :required, true
-            key :enum, %w(alexandria_convention)
             key :type, :string
           end
           parameter do
@@ -48,7 +43,6 @@ module Api
             key :in, :path
             key :description, 'Version of requested convention'
             key :required, true
-            key :enum, %w(latest 2.0.0 1.1.5 1.1.4 1.1.3)
             key :type, :string
           end
           parameter do
@@ -72,51 +66,31 @@ module Api
       end
 
       def load_schema
-        if File.exists?(@schema_pathname)
-          send_file @schema_pathname, type: @response_type, filename: @schema_filename
-        else
-          head 404 and return
-        end
-      end
-
-      private
-
-      def set_available_schemas
-        @schemas = {}
-        projects = Dir.entries(SCHEMAS_BASE_DIR).delete_if {|entry| entry.start_with?('.')}
-        projects.each do |project_name|
-          snapshots_path = SCHEMAS_BASE_DIR + "#{project_name}/snapshot"
-          snapshots = Dir.entries(snapshots_path).delete_if {|entry| entry.start_with?('.')}
-          versions = %w(latest) + Naturally.sort(snapshots).reverse
-          @schemas[project_name] = versions
-        end
-      end
-
-      def set_schema_filename
+        # set path to requested schema file
         project_name = params[:project_name]
         schema_format = params[:schema_format]
-        @schema_filename = "#{project_name}_schema.#{schema_format}"
-      end
-
-      def set_schema_filepath
-        project_name = params[:project_name]
         schema_version = params[:version]
-        @schema_pathname = SCHEMAS_BASE_DIR + project_name
+        schema_filename = "#{project_name}_schema.#{schema_format}"
+        schema_pathname = SCHEMAS_BASE_DIR + project_name
         if schema_version != 'latest'
-          @schema_pathname += "snapshot/#{params[:version]}"
+          schema_pathname += "snapshot/#{params[:version]}"
         end
-        @schema_pathname += @schema_filename
-      end
+        schema_pathname += schema_filename
 
-      def set_response_type
-        schema_format = params[:schema_format]
+        # determine response type
         case schema_format
         when 'json'
-          @response_type = 'application/json'
+          response_type = 'application/json'
         when 'tsv'
-          @response_type = 'text/plain'
+          response_type = 'text/plain'
         else
-          @response_type = 'application/json'
+          response_type = 'application/json'
+        end
+
+        if File.exists?(schema_pathname)
+          send_file schema_pathname, type: response_type, filename: schema_filename
+        else
+          head 404 and return
         end
       end
     end
